Cintas terminado. Ademas reparar el panel Peliculas y sus metodos CRUD panel Prestamos al 40%

+++ src/main/java/peaches/pelioficial/service/SocioService.java
@@ -45,7 +45,7 @@ public class SocioService {
         socioDAO.delete(socio);
     }
     
-    public Socio obtenerSocioId(long id){
+    public Socio obtenerSocioId(int id){
         Optional<Socio> socioOpt = socioDAO.get(id);
         return socioOpt.orElse(null);
     }
